testing changes

+++ Config/CONFIG00.h
@@ -18,12 +18,14 @@ namespace SENSORCONFIG
 
     inline int16_t conversion00( float const& voltage )
     {
-        return (voltage * 65.536 * 100.0);
+        // normalize across domain of 16 bit signed int
+        //(2^15-1)/500 = 65.536
+        return (int16_t)(voltage * 65.536 * 100.0);
     }
 
     inline int16_t conversion01( float const& voltage )
     {
-        return (voltage * 65.536 * 100.0);
+        return (int16_t)(voltage * 65.536 * 100.0);
     }
 
     // Sensor 0 Name
